refactor(core): move direct-room creation into the chat room helpers
`createOrGetDirectRoom` and the four functions it needs lived inside the
POST route module, so nothing outside an HTTP handler could open a direct
room without copying the direct-key rules. They move to `helpers.ts`, which
already owns `validateChatCoworkerIds` and the rest of the room vocabulary
the route imports.

Pure motion, no behaviour change. Separate from the caller that needs it so
this can be read as the move it is.
