+++ apps/core/src/routes/v1/chats/rooms/helpers.ts
@@ -15,11 +15,14 @@ import {
   publicChatRoomMessageMetadata,
   readUnfurlsFromMetadata,
 } from "@/helpers/chat-room-message-unfurl-metadata";
-import { badRequest, forbidden, notFound } from "@/helpers/error";
+import { badRequest, conflict, forbidden, notFound } from "@/helpers/error";
 import { resolveMemberOrganizationById } from "@/helpers/organization";
+import { isDirectKeyUniqueConstraintError } from "@/helpers/prisma";
 import prisma from "@/lib/db/prisma";
 import {
+  type ChatRoom,
   type ChatRoomMessageQuote,
+  chatRoomSchema,
   MAX_LISTED_CHAT_REACTION_REACTORS,
 } from "@/schemas/chat-room.schema";
 
@@ -2102,4 +2105,382 @@ export function resolveMentionedUserIds(params: {
   }
 
   return [...mentionedIds].filter((userId) => roomUserIds.has(userId));
+}
+
+async function findOrRestoreDirectByKey(
+  tx: {
+    chatRoom: Pick<typeof prisma.chatRoom, "findFirst" | "update">;
+  },
+  params: { organizationId: string | null; directKey: string },
+) {
+  const existing = await tx.chatRoom.findFirst({
+    where: {
+      organizationId: params.organizationId,
+      directKey: params.directKey,
+    },
+    include: chatRoomInclude,
+  });
+
+  if (!existing) {
+    return null;
+  }
+
+  if (!existing.archivedAt) {
+    return existing;
+  }
+
+  return await tx.chatRoom.update({
+    where: { id: existing.id },
+    data: { archivedAt: null },
+    include: chatRoomInclude,
+  });
+}
+async function serializeDirectRoomForViewer(
+  room: Parameters<typeof mapChatRoom>[0],
+  viewerUserId: string | null,
+  activeOrganizationId: string | null,
+): Promise<ChatRoom> {
+  return chatRoomSchema.parse(
+    viewerUserId
+      ? await mapChatRoomWithSidebarFlags(room, viewerUserId, prisma, {
+          activeOrganizationId,
+        })
+      : mapChatRoom(room),
+  );
+}
+function parseDirectCreateShape(params: {
+  currentUserId: string;
+  memberUserIds: readonly string[];
+  coworkerIds: readonly string[];
+}): DirectCreateShape {
+  const memberUserIds = normalizeUniqueStrings(params.memberUserIds);
+  const coworkerIds = normalizeUniqueStrings(params.coworkerIds);
+
+  if (memberUserIds.includes(params.currentUserId)) {
+    throw badRequest("Choose another organization member");
+  }
+
+  if (memberUserIds.length === 0 && coworkerIds.length === 0) {
+    throw badRequest("Choose a direct message target");
+  }
+
+  if (memberUserIds.length > 0 && coworkerIds.length > 0) {
+    throw badRequest("Group direct messages cannot include coworkers.");
+  }
+
+  if (coworkerIds.length > 1) {
+    throw badRequest("Direct messages support one coworker only.");
+  }
+
+  if (memberUserIds.length >= 1 && coworkerIds.length === 0) {
+    return {
+      kind: "human-direct",
+      memberUserIds,
+      coworkerIds: [],
+    };
+  }
+
+  if (memberUserIds.length === 0 && coworkerIds.length === 1) {
+    return {
+      kind: "coworker-1to1",
+      memberUserIds: [],
+      coworkerIds: [coworkerIds[0]],
+    };
+  }
+
+  throw badRequest("Choose a direct message target");
+}
+/**
+ * Direct rooms are addressed by their participant set, so creation is
+ * create-or-get: two clients opening the same conversation must land on one
+ * room instead of racing into duplicates.
+ *
+ * Coworker 1:1 inherits the active organization when set, else personal.
+ * Human 1:1 reuses a Personal Direct if one exists, else an Org Direct when
+ * both are Members of the active org, else creates a Personal Direct when
+ * they share an External channel. Multi-human groups stay org-scoped.
+ */
+/**
+ * Valid direct create targets: human-direct (≥1 humans, no coworkers) or
+ * coworker-1to1 (exactly one coworker, no humans). Mix / multi-coworker /
+ * empty are invalid.
+ */
+type DirectCreateShape =
+  | {
+      kind: "human-direct";
+      memberUserIds: string[];
+      coworkerIds: [];
+    }
+  | {
+      kind: "coworker-1to1";
+      memberUserIds: [];
+      coworkerIds: [string];
+    };
+
+export async function createOrGetDirectRoom(params: {
+  organizationId: string | null;
+  currentUserId: string;
+  memberUserIds: readonly string[];
+  coworkerIds: readonly string[];
+  /** Sidebar viewer. Null skips pin/mute/unread (coworker actor has none). */
+  viewerUserId?: string | null;
+}): Promise<{ room: ChatRoom; created: boolean }> {
+  const { currentUserId } = params;
+  const viewerUserId =
+    params.viewerUserId === undefined ? currentUserId : params.viewerUserId;
+  const shape = parseDirectCreateShape({
+    currentUserId,
+    memberUserIds: params.memberUserIds,
+    coworkerIds: params.coworkerIds,
+  });
+  const requestedMemberUserIds = shape.memberUserIds;
+  const requestedCoworkerIds = shape.coworkerIds;
+  const activeOrganizationId = params.organizationId;
+
+  // Holds the key computed inside the transaction so a directKey race
+  // can reuse it; a plain `let` would be narrowed to `never` by control flow
+  // analysis because the assignment happens inside the callback.
+  const directKeyRef: { current: string | null } = { current: null };
+  const createOrganizationIdRef: { current: string | null } = {
+    current: activeOrganizationId,
+  };
+
+  try {
+    const result = await prisma.$transaction(async (tx) => {
+      if (activeOrganizationId) {
+        if (shape.kind === "human-direct") {
+          await resolveMemberOrganizationById({
+            id: activeOrganizationId,
+            userId: currentUserId,
+            tx,
+          });
+        } else {
+          // Coworker 1:1 owner is the human on the room (user actor or
+          // originated target). A missing owner is a bad target, not
+          // "you are not a member".
+          await validateOrganizationUserIds(
+            activeOrganizationId,
+            [currentUserId],
+            tx,
+          );
+        }
+      }
+
+      if (shape.kind === "coworker-1to1") {
+        const workspaceId = await resolveWorkspaceIdForChatRoom({
+          organizationId: activeOrganizationId,
+          personalUserId: currentUserId,
+          tx,
+        });
+        const coworkerIds = await validateChatCoworkerIds(
+          requestedCoworkerIds,
+          workspaceId,
+          tx,
+        );
+        const directKey = buildDirectParticipantRoomKey({
+          currentUserId,
+          memberUserIds: [],
+          coworkerIds,
+        });
+        directKeyRef.current = directKey;
+        createOrganizationIdRef.current = activeOrganizationId;
+
+        const existing = await findOrRestoreDirectByKey(tx, {
+          organizationId: activeOrganizationId,
+          directKey,
+        });
+        if (existing) {
+          return { room: existing, created: false };
+        }
+
+        return createDirectRoomRecord({
+          tx,
+          currentUserId,
+          organizationId: activeOrganizationId,
+          directKey,
+          memberUserIds: [],
+          coworkerIds,
+        });
+      }
+
+      const isGroup = requestedMemberUserIds.length > 1;
+      const orgTeammateIds = activeOrganizationId
+        ? await filterOrganizationUserIds(
+            activeOrganizationId,
+            requestedMemberUserIds,
+            tx,
+          )
+        : [];
+      const targetsAreOrgTeammates =
+        activeOrganizationId != null &&
+        orgTeammateIds.length === requestedMemberUserIds.length;
+
+      if (isGroup && !activeOrganizationId) {
+        throw badRequest("Switch to an organization to message a teammate.");
+      }
+      if (isGroup && !targetsAreOrgTeammates) {
+        throw badRequest("Room human members must belong to the organization");
+      }
+
+      const memberUserIds = requestedMemberUserIds;
+      const directKey = buildDirectParticipantRoomKey({
+        currentUserId,
+        memberUserIds,
+        coworkerIds: [],
+      });
+      directKeyRef.current = directKey;
+
+      const existing = await findLiveDirectByParticipantKey(
+        tx,
+        directKey,
+        targetsAreOrgTeammates ? activeOrganizationId : null,
+      );
+      if (existing) {
+        return { room: existing, created: false };
+      }
+
+      if (!targetsAreOrgTeammates) {
+        const peerUserId = memberUserIds[0];
+        if (!peerUserId) {
+          throw badRequest("Choose a direct message target");
+        }
+        const shareChannel = await usersShareExternalChannel(
+          currentUserId,
+          peerUserId,
+          tx,
+        );
+        if (!shareChannel) {
+          throw badRequest(
+            "You can only message people you share an external channel with.",
+          );
+        }
+      }
+
+      const organizationId = targetsAreOrgTeammates
+        ? activeOrganizationId
+        : null;
+      createOrganizationIdRef.current = organizationId;
+
+      return createDirectRoomRecord({
+        tx,
+        currentUserId,
+        organizationId,
+        directKey,
+        memberUserIds,
+        coworkerIds: [],
+      });
+    });
+
+    return {
+      room: await serializeDirectRoomForViewer(
+        result.room,
+        viewerUserId,
+        activeOrganizationId,
+      ),
+      created: result.created,
+    };
+  } catch (error) {
+    // directKey race: another request won the create — return that room.
+    if (isDirectKeyUniqueConstraintError(error) && directKeyRef.current) {
+      const existing =
+        shape.kind === "coworker-1to1"
+          ? await findOrRestoreDirectByKey(prisma, {
+              organizationId: createOrganizationIdRef.current,
+              directKey: directKeyRef.current,
+            })
+          : await findLiveDirectByParticipantKey(
+              prisma,
+              directKeyRef.current,
+              createOrganizationIdRef.current,
+            );
+
+      if (existing) {
+        return {
+          room: await serializeDirectRoomForViewer(
+            existing,
+            viewerUserId,
+            activeOrganizationId,
+          ),
+          created: false,
+        };
+      }
+
+      throw conflict("Direct room already exists");
+    }
+
+    throw error;
+  }
+}
+async function createDirectRoomRecord(params: {
+  tx: Prisma.TransactionClient;
+  currentUserId: string;
+  organizationId: string | null;
+  directKey: string;
+  memberUserIds: readonly string[];
+  coworkerIds: readonly string[];
+}) {
+  const {
+    tx,
+    currentUserId,
+    organizationId,
+    directKey,
+    memberUserIds,
+    coworkerIds,
+  } = params;
+
+  const [targetUsers, targetCoworkers] = await Promise.all([
+    memberUserIds.length > 0
+      ? tx.user.findMany({
+          where: { id: { in: [...memberUserIds] } },
+          select: { id: true, name: true, email: true },
+        })
+      : Promise.resolve([]),
+    coworkerIds.length > 0
+      ? tx.coworker.findMany({
+          where: { id: { in: [...coworkerIds] } },
+          select: { id: true, name: true },
+        })
+      : Promise.resolve([]),
+  ]);
+  const usersById = new Map(targetUsers.map((user) => [user.id, user]));
+  const coworkersById = new Map(
+    targetCoworkers.map((coworker) => [coworker.id, coworker]),
+  );
+  const directName = buildDirectRoomName([
+    ...memberUserIds.map((userId) => {
+      const user = usersById.get(userId);
+      return user?.name || user?.email || userId;
+    }),
+    ...coworkerIds.map((coworkerId) => {
+      return coworkersById.get(coworkerId)?.name || coworkerId;
+    }),
+  ]);
+  const room = await tx.chatRoom.create({
+    data: {
+      organizationId,
+      createdByUserId: currentUserId,
+      name: directName,
+      slug: null,
+      kind: "direct",
+      directKey,
+      userMembers: {
+        create: [
+          { userId: currentUserId },
+          ...memberUserIds.map((userId) => ({ userId })),
+        ],
+      },
+      readStates: {
+        create: [
+          { userId: currentUserId },
+          ...memberUserIds.map((userId) => ({ userId })),
+        ],
+      },
+      coworkerMembers: {
+        create: coworkerIds.map((coworkerId) => ({ coworkerId })),
+      },
+    },
+    include: chatRoomInclude,
+  });
+
+  return { room, created: true };
 }
